Add files via upload
This is a Monte Carlo simulation that creates randomly generated numbers that output meaningful probability densities. This action is required especially in particle physics where engineered instruments are not qualified to measure tiny differences. For example in test.pdf, it can be easily seen that kink angles for the specified particles are quite small and there is no instrument that can measure such distances, especially the particle Ds which is an important particle in tau neutrino research where it contributes more than %90 of the tau neutrino creation.

Code is mainly a functional code that consists of many functions each specified to their own filler value but the histogram drawing process is automated as much as possible where they intersect with each other.

Automated code is the Main.py that imports functions from MyFunctions.py. A non-automated version of the Main.py can be seen from the Hardcoded.py. This is a code that is written with Python using the Cern Root framework which name is PyRoot. This framework enables the analysis of a large number of data set.